requirements.txt
requests
python-dotenv
Pillow
tqdm
discord.py==2.3.2
supabase==2.3.4
aiohttp
beautifulsoup4
loguru
